Allow setting the log level through Logging.set_logger
As part of the introspection changes the method signature was changed
from accepting a path, log level, and rotation frequency to accepting a
path, rotation frequency and rotation size. This was done without
updating any client applications to use the new signature, fortunately
we've only got one application under active use that has this library
included.

Change-Id: Ic8c0ea55847043a8f3644425b31cca41e84c1fb5
Refs: SIS-835