Stop double logging in development mode
In development mode Rails wants to make double sure we're logging stuff
to STDERR. When combined with our defaulting to STDOUT this causes us
to get everything to the console twice. There are two ways to stop this
behavior:

1. Make ActiveSupport::Logger.broadcast return an empty module,
   effectively removing the functionality while preserving the API. This
   will break any applications that use the broadcast functionality.
2. Remove the console block from the ActiveRecord Railtie preventing it
   from doing the extra console behaviors. The downside of this approach
   being that it would break the console sandbox functionality.

The choice for #1 was made as the potential bugs it introduces are not
going to potentially bite someone in a production console changing data
they don't really intend to.

Change-Id: I6a5e8f5163fcefbb93600c54eeb55d0a20dfe2cf
Refs: SIS-1394